test: move integration tests to existing IT class
That way, surefire will only run non-containerized tests.
