+++ src/test/scala/org/neo4j/spark/util/ValidationsIT.scala
@@ -16,6 +16,7 @@
  */
 package org.neo4j.spark.util
 
+import org.apache.spark.sql.SparkSession
 import org.junit.jupiter.api.Assertions._
 import org.junit.jupiter.api.Test
 import org.neo4j.driver.AccessMode
@@ -179,6 +180,54 @@ class ValidationsIT extends SparkConnectorScalaSuiteIT {
         s"Invalid query `$query` because the accepted types are [WRITE_ONLY, READ_WRITE], but the actual type is READ_ONLY"
       )
     )
+  }
+
+  @Test
+  def testVersionThrowsExceptionSparkVersionIsNotSupported(): Unit = {
+    val sparkVersion = SparkSession.getActiveSession
+      .map { _.version }
+      .getOrElse("UNKNOWN")
+    try {
+      Validations.validate(ValidateSparkMinVersion("4.10000"))
+      fail(s"should be thrown a ${classOf[IllegalArgumentException].getName}")
+    } catch {
+      case e: IllegalArgumentException =>
+        assertEquals(
+          s"""Your current Spark version $sparkVersion is not supported by the current connector.
+             |Please visit https://neo4j.com/developer/spark/overview/#_spark_compatibility to know which connector version you need.
+             |""".stripMargin,
+          e.getMessage
+        )
+      case e: Throwable =>
+        fail(s"should be thrown a ${classOf[IllegalArgumentException].getName}, got ${e.getClass} instead")
+    }
+  }
+
+  @Test
+  def testVersionShouldBeValid(): Unit = {
+    val fullVersion = SparkSession
+      .getDefaultSession
+      .map(_.version)
+      .getOrElse("3.2")
+    val baseVersion = fullVersion
+      .split("\\.")
+      .take(2)
+      .mkString(".")
+    Validations.validate(ValidateSparkMinVersion(s"$baseVersion.*"))
+    Validations.validate(ValidateSparkMinVersion(fullVersion))
+    Validations.validate(ValidateSparkMinVersion(s"$fullVersion-amzn-0"))
+  }
+
+  @Test
+  def testVersionShouldValidateTheVersion(): Unit = {
+    val version = ValidateSparkMinVersion("2.3.0")
+    assertTrue(version.isSupported("2.3.0-amzn-1"))
+    assertTrue(version.isSupported("2.3.1-amzn-1"))
+    assertTrue(version.isSupported("3.3.0-amzn-1"))
+    assertTrue(version.isSupported("3.3.0"))
+    assertTrue(version.isSupported("3.1.0"))
+    assertTrue(version.isSupported("3.2.0"))
+    assertFalse(version.isSupported("2.2.10"))
   }
 
 }
